refactor(core): move background tool runs out of the kernel
`spawn_background` lived in `everruns-core` while the capability advertising
it lived in platform. The tool creates session tasks and schedules, mirrors a
detached run's lifecycle onto a task, and streams output through the session's
event sink — hosted behaviour backed by hosted services, not neutral kernel
execution.

Moves the whole runtime to `everruns-platform` as `background_run`: the tool
and its schedule path, the session-task mirroring, the background event sink,
the worker- and session-scoped admission permits, and the reattach entry point
used after worker loss. The `background_execution` capability now owns the tool
it advertises, and `subagents` shares the same permits so every detached path
goes through one gate. Core keeps the neutral `BackgroundExecutableTool` and
`BackgroundEventSink` contracts in `core::background` and runs whatever a host
supplies. Tests and fixtures moved with the code; core's tools.rs drops from
3839 to ~1300 lines.

First slice of EVE-888. It does NOT free the session_task record for EVE-880,
contrary to the expectation recorded against EVE-897: `SessionTask` and
`TaskMessage` are embedded in the canonical task.created / task.updated /
task.message.* event payloads, which EVE-888 retains as kernel surface while
putting canonical event semantics out of scope. `wake_queue` and
`task_observer` consume the record too. Moving that family needs a neutral task
projection for events or an accepted payload change — a decision, not a
mechanical move. Recorded in the knowledge log.

Refs EVE-888
